+++ src/Cards.cpp
@@ -105,7 +105,7 @@ std::vector<Card*> Hand::getCardsOnHand() {
 // Stream overloading for Hand.
 std::ostream& operator<<(std::ostream &os, Hand &hand) {
     os << "There are " << hand.getCardsOnHand().size() << " cards on Hand:" << std::endl;
-    for(int i = 0; i < hand.getCardsOnHand().size(); i++) {
+    for(size_t i = 0; i < hand.getCardsOnHand().size(); i++) {
         os << "  Index " << i << ": " << hand.getCardsOnHand().at(i)->getCard() << std::endl;
     }
     return os;
@@ -133,7 +133,7 @@ void Hand::showHand(Hand& hand) {
 
 // Destructor for Hand and deletes the Card* of each card in Hand.
 Hand::~Hand() {
-    for(int i = 0; i < cardsOnHand.size(); i++) {
+    for(size_t i = 0; i < cardsOnHand.size(); i++) {
         delete cardsOnHand.at(i);
     }
 }
@@ -143,7 +143,7 @@ Deck::Deck() : cardsOnDeck() {}
 
 // Deep copy constructor for Deck.
 Deck::Deck(Deck const &deck){
-    for(int i = 0; i < deck.cardsOnDeck.size(); i++) {
+    for(size_t i = 0; i < deck.cardsOnDeck.size(); i++) {
         Card* cardptr = deck.cardsOnDeck.at(i);
         if (cardptr) {
             Card* cardCopied = new Card(cardptr->getCard());
@@ -173,7 +173,7 @@ std::vector<Card*> Deck::getCardsOnDeck() {
 std::ostream & operator << (std::ostream &os, Deck &deck) {
     os << "There are " << deck.getCardsOnDeck().size() << " cards on the Deck:" << std::endl;
     
-    for(int i = 0; i < deck.getCardsOnDeck().size(); i++) {
+    for(size_t i = 0; i < deck.getCardsOnDeck().size(); i++) {
         os << "  " << deck.getCardsOnDeck().at(i)->getCard() << std::endl;
     }
 
@@ -224,7 +224,7 @@ void Deck::showDeck(Deck &deck) {
 
 // Destructor for Deck. Deletes the Card* of each card in the Deck.
 Deck::~Deck() {
-    for(int i = 0; i < cardsOnDeck.size(); i++) {
+    for(size_t i = 0; i < cardsOnDeck.size(); i++) {
         delete cardsOnDeck.at(i);
     }
 }
